ref: moved ApiFilterProps to /schemas/filters
