+++ functions/orders/getOrders.ts
@@ -1,6 +1,6 @@
 import { useWoocommerce } from '@/api/useWoocommerce'
 import { Order, validateOrders } from '@/schemas/orders/Order'
-import { ApiFilterProps, validateApiFilterProps } from '@/schemas/ApiFilterProps'
+import { ApiFilterProps, validateApiFilterProps } from '@/schemas/filters/ApiFilterProps'
 
 export async function getOrders(filterProps: ApiFilterProps = {}): Promise<Order[] | never> {
   const filters = validateApiFilterProps(filterProps)
