+++ functions/products/getProducts.ts
@@ -1,6 +1,6 @@
 import { useWoocommerce } from '@/api/useWoocommerce'
 import { Product, validateProducts } from '@/schemas/products/Product'
-import { ApiFilterProps, validateApiFilterProps } from '@/schemas/ApiFilterProps'
+import { ApiFilterProps, validateApiFilterProps } from '@/schemas/filters/ApiFilterProps'
 
 export async function getProducts(filterProps: ApiFilterProps): Promise<Product[] | never> {
   const filters = validateApiFilterProps(filterProps)
